Minor tweaks to the look and feel of the main window.
Redesigned the "connection string helper", adding a tab control for different types, and adjusting the wording of the help text.

Added some error handling throughout the project - some errors are ignored, which I realize is not very good practice, but I have limited time to work on this project, and likely not many people will ever use it. If it becomes a problem, I will correct it.